[22.07.19] 오후 paging 기능 구현 완료

+++ src/main/java/com/project/web_prj/board/repository/BoardRepository.java
@@ -32,7 +32,7 @@ public interface BoardRepository {
 
 
     // 전체 게시물 수 조회
-    Long getTotalCount();
+    int getTotalCount();
 
 
     // 조회수 상승 처리
